[COMMON] g2d: fix unintended sign extension
Multiplying crop_w and crop_h could be promoted in
type signed 32bit, then signed extended to type
unsigned long long, so fix signed extension.

(cherry picked from commit 66c8b9c114091b73fd960b3257b7df285ac6f1c4)
Bug: 165562344

+++ drivers/gpu/exynos/g2d/g2d_perf.c
@@ -72,7 +72,7 @@ static u32 perf_basis[PPC_SC] = {1024, 1023, 256, 113, 64, 0};
 static char perf_index_sc(struct g2d_performance_layer_data *layer)
 {
 	u32 ratio = (((u64)layer->window_w * layer->window_h) << 10) /
-			(layer->crop_w * layer->crop_h);
+			((u32)layer->crop_w * layer->crop_h);
 	int i;
 
 	for (i = 0; i < PPC_SC; i++) {
